requirements.txt

python-telegram-bot==20.7
docx2pdf
Flask
